차트 및 인사이트 관련 개선 및 Clova Studio 프롬프트 개선(감정등록 및 요약 관련)

+++ src/main/java/com/melog/melog/emotion/domain/model/response/EmotionChartResponse.java
@@ -8,6 +8,6 @@ import java.util.Map;
 @Getter
 @Builder
 public class EmotionChartResponse {
-    private Map<String, Integer> thisMonth;
-    private Map<String, Integer> compareWithLastMonth;
+    private Map<String, Double> thisMonth;
+    private Map<String, Double> compareWithLastMonth;
 }
